+++ src/main/java/com/example/yes48/Service/AdminService.java
@@ -1,10 +1,12 @@
 package com.example.yes48.Service;
 
+import com.example.yes48.domain.FileStore;
 import com.example.yes48.domain.goods.Goods;
 import com.example.yes48.repository.AdminRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.List;
@@ -15,8 +17,6 @@ public class AdminService {
 
     @Autowired
     private AdminRepository goodsRepository;
-    @Autowired
-    FileStoreService fileStoreService;
 
     /**
      * 상품 등록
